+++ sdk/openai/Azure.AI.OpenAI/src/Custom/ChatCompletions/ChatMessageImageContentItem.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
+using System.IO;
 using System.Net.Mime;
 using Azure.Core;
 
@@ -41,60 +42,12 @@ public partial class ChatMessageImageContentItem : ChatMessageContentItem
 
     /// <summary> Initializes a new instance of ChatMessageImageContentItem using default detail. </summary>
     /// <param name="imageUri"> An internet location, which must be accessible to the model,from which the image may be retrieved. </param>
-    /// <exception cref="ArgumentNullException"> <paramref name="imageUri"/> is null. </exception>
-    public ChatMessageImageContentItem(Uri imageUri)
-        : this(new InternalChatMessageImageUrl(imageUri.AbsoluteUri))
-    {
-        ImageUri = imageUri;
-    }
-
-    /// <summary> Initializes a new instance of ChatMessageImageContentItem. </summary>
-    /// <param name="imageUri"> An internet location, which must be accessible to the model,from which the image may be retrieved. </param>
     /// <param name="detailLevel"> The image detail level the model should use when evaluating the image. </param>
     /// <exception cref="ArgumentNullException"> <paramref name="imageUri"/> is null. </exception>
-    public ChatMessageImageContentItem(Uri imageUri, ChatMessageImageDetailLevel detailLevel)
+    public ChatMessageImageContentItem(Uri imageUri, ChatMessageImageDetailLevel? detailLevel = null)
         : this(new InternalChatMessageImageUrl(imageUri.AbsoluteUri, detailLevel))
     {
         ImageUri = imageUri;
-    }
-
-    /// <summary>
-    /// Initializes a new instance of <see cref="ChatMessageImageContentItem"/> using JPEG image data and the
-    /// default <see cref="ChatMessageImageDetailLevel"/> setting.
-    /// </summary>
-    /// <param name="jpegImageData"> The image data to provide to the model. </param>
-    public ChatMessageImageContentItem(BinaryData jpegImageData)
-        : this(new InternalChatMessageImageUrl(jpegImageData))
-    {
-        ImageData = jpegImageData;
-        ImageDataFormat = MediaTypeNames.Image.Jpeg;
-    }
-
-    /// <summary>
-    /// Initializes a new instance of <see cref="ChatMessageImageContentItem"/> using JPEG image data and a
-    /// specified <see cref="ChatMessageImageDetailLevel"/> setting.
-    /// </summary>
-    /// <param name="jpegImageData"> The image data to provide to the model. </param>
-    /// <param name="detailLevel"> The detail level to use when processing the provided image data. </param>
-    public ChatMessageImageContentItem(BinaryData jpegImageData, ChatMessageImageDetailLevel detailLevel)
-        : this(new InternalChatMessageImageUrl(jpegImageData, detailLevel))
-    {
-        ImageData = jpegImageData;
-        ImageDataFormat = MediaTypeNames.Image.Jpeg;
-        // detailLevel is retaiend via InternalChatMessageImageUrl
-    }
-
-    /// <summary>
-    /// Initializes a new instance of <see cref="ChatMessageImageContentItem"/> using image data in a specified format
-    /// and the default <see cref="ChatMessageImageDetailLevel"/> setting.
-    /// </summary>
-    /// <param name="imageData"> The image data to provide to the model. </param>
-    /// <param name="imageDataFormat"> The image format string, e.g. 'image/jpeg', associated with the data. </param>
-    public ChatMessageImageContentItem(BinaryData imageData, string imageDataFormat)
-        : this(new InternalChatMessageImageUrl(imageData, imageDataFormat))
-    {
-        ImageData = imageData;
-        ImageDataFormat = imageDataFormat;
     }
 
     /// <summary>
@@ -104,12 +57,52 @@ public partial class ChatMessageImageContentItem : ChatMessageContentItem
     /// <param name="imageData"> The image data to provide to the model. </param>
     /// <param name="imageDataFormat"> The image format string, e.g. 'image/jpeg', associated with the data. </param>
     /// <param name="detailLevel"> The detail level to use when processing the provided image data. </param>
-    public ChatMessageImageContentItem(BinaryData imageData, string imageDataFormat, ChatMessageImageDetailLevel detailLevel)
+    public ChatMessageImageContentItem(BinaryData imageData, string imageDataFormat, ChatMessageImageDetailLevel? detailLevel = null)
         : this(new InternalChatMessageImageUrl(imageData, imageDataFormat, detailLevel))
     {
         ImageData = imageData;
         ImageDataFormat = imageDataFormat;
         // detailLevel is retaiend via InternalChatMessageImageUrl
+    }
+
+    /// <summary>
+    /// Initializes a new instance of <see cref="ChatMessageImageContentItem"/> using image data from a provided local
+    /// file path. Image format is inferred from the file extension.
+    /// </summary>
+    /// <param name="localImageFilePath">
+    ///     The path to a locally accessible image file with a file extension that accurately represents its format.
+    /// </param>
+    /// <param name="detailLevel"> The detail level to use when processing the provided image data. </param>
+    public ChatMessageImageContentItem(string localImageFilePath, ChatMessageImageDetailLevel? detailLevel = null)
+    {
+        Argument.AssertNotNull(localImageFilePath, nameof(localImageFilePath));
+        FileInfo localFileInfo = new(localImageFilePath);
+        using Stream localFileStream = localFileInfo.OpenRead();
+        BinaryData imageData = BinaryData.FromStream(localFileStream);
+        string imageDataFormat = GetImageFormatFromExtension(localFileInfo.Extension);
+        ImageData = imageData;
+        ImageDataFormat = imageDataFormat;
+        InternalImageUrlOrData = new(imageData, imageDataFormat, detailLevel);
+    }
+
+    /// <summary>
+    /// Initializes a new instance of <see cref="ChatMessageImageContentItem"/> using image data from a provided local
+    /// file path with a specific image format.
+    /// </summary>
+    /// <param name="localImageFilePath">
+    ///     The path to a locally accessible image file with a file extension that accurately represents its format.
+    /// </param>
+    /// <param name="imageDataFormat"> The format, e.g. 'image/png', of the provided image file. </param>
+    /// <param name="detailLevel"> The detail level to use when processing the provided image data. </param>
+    public ChatMessageImageContentItem(string localImageFilePath, string imageDataFormat, ChatMessageImageDetailLevel? detailLevel = null)
+    {
+        Argument.AssertNotNull(localImageFilePath, nameof(localImageFilePath));
+        FileInfo localFileInfo = new(localImageFilePath);
+        using Stream localFileStream = localFileInfo.OpenRead();
+        BinaryData imageData = BinaryData.FromStream(localFileStream);
+        ImageData = imageData;
+        ImageDataFormat = imageDataFormat;
+        InternalImageUrlOrData = new(imageData, imageDataFormat, detailLevel);
     }
 
     /// <summary> Initializes a new instance of <see cref="ChatMessageImageContentItem"/>. </summary>
@@ -129,5 +122,18 @@ public partial class ChatMessageImageContentItem : ChatMessageContentItem
     internal ChatMessageImageContentItem(string type, InternalChatMessageImageUrl imageUrl) : base(type)
     {
         InternalImageUrlOrData = imageUrl;
+    }
+
+    private static string GetImageFormatFromExtension(string imageFileExtension)
+    {
+        string extension = imageFileExtension?.ToLowerInvariant();
+        string imageDataFormat = extension switch
+        {
+            ".jpg" or ".jpeg" => MediaTypeNames.Image.Jpeg,
+            { } when extension?.Length > 0 => $"image/{extension.Substring(1)}",
+            _ => throw new ArgumentException(
+                $"Unable to infer image file format from provided file extension: {extension}"),
+        };
+        return imageDataFormat;
     }
 }